requirements.txt
discord.py==1.6.0
virtualenv==20.4.2